add ruff formatting to pre-commit
