+++ pyproject.toml
@@ -22,6 +22,17 @@ authors = [{ name = "Jose Morales", email = "jmoralz92@gmail.com" }]
 readme = "README.md"
 keywords = ["forecasting", "time-series"]
 
+[project.optional-dependencies]
+dev = [
+    "numba",
+    "pandas",
+    "pre-commit",
+    "pytest",
+    "ruff",
+    "utilsforecast",
+    "window-ops",
+]
+
 [project.urls]
 homepage = "https://nixtla.github.io/coreforecast"
 documentation = "https://nixtla.github.io/coreforecast"
@@ -44,3 +55,10 @@ wheel.py-api = "py3"
 archs = "all"
 build-verbosity = 3
 test-command = 'python -c "import coreforecast._lib"'
+
+[tool.pyright]
+venvPath = "."
+venv = ".venv"
+
+[tool.ruff.lint]
+select = ["F", "I"]
