+++ tests/__init__.py
@@ -1,0 +1,52 @@
+import numpy as np
+from coreforecast.scalers import (
+    LocalMinMaxScaler,
+    LocalRobustScaler,
+    LocalStandardScaler,
+)
+
+
+def std_scaler_stats(x):
+    return np.nanmean(x), np.nanstd(x)
+
+
+def minmax_scaler_stats(x):
+    min, max = np.nanmin(x), np.nanmax(x)
+    return min, max - min
+
+
+def robust_scaler_iqr_stats(x):
+    q25, median, q75 = np.nanquantile(x, [0.25, 0.5, 0.75])
+    return median, q75 - q25
+
+
+def robust_scaler_mad_stats(x):
+    median = np.nanmedian(x)
+    mad = np.nanmedian(np.abs(x - median))
+    return median, mad
+
+
+def scaler_transform(x, stats):
+    offset, scale = stats
+    return (x - offset) / scale
+
+
+def scaler_inverse_transform(x, stats):
+    offset, scale = stats
+    return x * scale + offset
+
+
+scaler2fns = {
+    "standard": std_scaler_stats,
+    "minmax": minmax_scaler_stats,
+    "robust-iqr": robust_scaler_iqr_stats,
+    "robust-mad": robust_scaler_mad_stats,
+}
+scaler2core = {
+    "standard": LocalStandardScaler(),
+    "minmax": LocalMinMaxScaler(),
+    "robust-iqr": LocalRobustScaler("iqr"),
+    "robust-mad": LocalRobustScaler("mad"),
+}
+scalers = list(scaler2fns.keys())
+dtypes = [np.float32, np.float64]
